fix: Move SHA512_256 to testing ONLY, outsource Base32 to package
